requirements.txt
# ============================================
# Core Dependencies
# ============================================
web3==7.6.0
python-dotenv==1.0.1
requests==2.31.0

# ============================================
# Data Processing & ML
# ============================================
pandas==2.2.0
numpy==1.26.4
arch==6.3.0

# ============================================
# Database
# ============================================
supabase==2.3.4

# ============================================
# Development & Testing
# ============================================
pytest==8.0.0
flake8==7.1.1
black==24.2.0

# ============================================
# Production Monitoring (Optional)
# ============================================
# sentry-sdk==1.40.0

